chore: update project dependencies and refactor voice_generator module
Modified `pyproject.toml` to include optional dependencies for `gradio` and added new packages in `uv.lock`. Refactored the `voice_generator` module by consolidating core constants and environment checks into `__main__.py`, improving code organization and clarity.
